+++ CHANGES.md
@@ -4,6 +4,7 @@ I will try to keep track of changes between releases here
 
 ## Upcomming release, available in devel branch
 
+- preventing iSH-AOK from using Alpine 3.20
 - Alpine & Debin update cmds
 - debian_apt_update & alpine_apk_update
 - do_chorot - show details for colliding pids